refactor(web): split FOGPage render/POST helpers into two traits
Move the 16 reusable page-construction methods out of FOGPage into two
filename-loaded traits, byte-for-byte, so the controller base stops
growing without bound (5810 -> 4921 lines).

- FOGPageRender: the GET/read surface — create/edit form builders,
  association/history/edit tab renderers, and the AJAX read endpoints
  that feed their tables.
- FOGPagePost: the POST/write surface — the create/edit/association
  POST orchestrators, schedule-type validation, and the JSON-response
  helpers they answer with.

The render/POST boundary mirrors the creation->handleAddPost /
edit->handleEditPost tracing split: a render bug and a submit bug land
in different files. Behavior is identical because a trait's methods
compile into the using class as if declared there (same $this, same
access to inherited statics like self::$HookManager), so every existing
call site resolves unchanged. The files are named <Trait>.class.php so
the existing filename-keyed autoloader resolves `use` with no
autoloader change.
