+++ packages/web/lib/fog/fogpage.class.php
@@ -23,6 +23,8 @@
  */
 abstract class FOGPage extends FOGBase
 {
+    use FOGPageRender;
+    use FOGPagePost;
     /**
      * Name of the page
      *
@@ -1452,203 +1454,6 @@ abstract class FOGPage extends FOGBase
         return ob_get_clean();
     }
     /**
-     * Emits a JSON response body and terminates the request.
-     *
-     * This is the universal terminal shared by every AJAX endpoint:
-     * set the HTTP status, echo the (already-encoded) body, and exit.
-     * Callers that fire a hook must do so before calling this; use
-     * jsonHookResponse() when a result hook needs to mutate the body.
-     *
-     * @param int    $code The HTTP status code to send.
-     * @param string $body The response body (already JSON-encoded).
-     *
-     * @return void
-     */
-    protected static function jsonSend($code, $body)
-    {
-        http_response_code($code);
-        echo $body;
-        exit;
-    }
-    /**
-     * Fires a result hook then emits the JSON response.
-     *
-     * Preserves the existing per-method hook contract exactly: the
-     * caller passes the same by-reference argument array it always
-     * has (including 'code' and 'msg'), so plugins registered on
-     * $hook can still mutate the status code and body. The (possibly
-     * mutated) values are read back from that same array after the
-     * event fires. PHP preserves the member references when the array
-     * is passed by value, so $args['code']/$args['msg'] resolve to the
-     * caller's $code/$msg exactly as the inline code did.
-     *
-     * @param array  $args The by-reference hook argument array; must
-     *                      contain 'code' and 'msg' keys.
-     * @param string $hook The hook event name to fire.
-     *
-     * @return void
-     */
-    protected function jsonHookResponse(array $args, $hook)
-    {
-        self::$HookManager->processEvent($hook, $args);
-        $this->jsonSend($args['code'], $args['msg']);
-    }
-    /**
-     * Shared scaffold for the create (addPost) AJAX handlers.
-     *
-     * Owns the boilerplate every create endpoint repeated verbatim:
-     * the auth/CSRF gate, the JSON content-type header, the
-     * "<BASE>_POST" pre-event, the $serverFault flag, the try/catch
-     * that turns a thrown Exception into the proper HTTP status, the
-     * success/fail hook names and JSON body, and the terminal
-     * jsonHookResponse() that fires the result hook and emits the
-     * response.
-     *
-     * The page-specific part — reading $_POST, validating, building
-     * and saving the entity — lives in the $build closure. The closure
-     * receives $serverFault by reference (set it true before throwing
-     * to signal an HTTP 500 rather than a 400) and must return the
-     * saved entity, which is handed to the result hook under
-     * $entityKey so listeners registered on "<BASE>_SUCCESS" /
-     * "<BASE>_FAIL" still see it exactly as before.
-     *
-     * @param string   $entityKey    Payload key for the entity (e.g. 'Group').
-     * @param string   $hookBase     Hook prefix (e.g. 'GROUP_ADD'); the
-     *                               _POST/_SUCCESS/_FAIL events derive from it.
-     * @param string   $successMsg   Translated success message body.
-     * @param string   $successTitle Translated success title.
-     * @param string   $failTitle    Translated failure title.
-     * @param callable $build        Closure(&$serverFault): returns the entity.
-     *
-     * @return void
-     */
-    protected function handleAddPost(
-        $entityKey,
-        $hookBase,
-        $successMsg,
-        $successTitle,
-        $failTitle,
-        callable $build
-    ) {
-        self::checkAuthAndCSRF();
-        header('Content-type: application/json');
-        self::$HookManager->processEvent($hookBase . '_POST');
-        $serverFault = false;
-        $Entity = null;
-        try {
-            $Entity = $build($serverFault);
-            $code = HTTPResponseCodes::HTTP_CREATED;
-            $hook = $hookBase . '_SUCCESS';
-            $msg = json_encode(
-                [
-                    'msg' => $successMsg,
-                    'title' => $successTitle
-                ]
-            );
-        } catch (Exception $e) {
-            $code = (
-                $serverFault ?
-                HTTPResponseCodes::HTTP_INTERNAL_SERVER_ERROR :
-                HTTPResponseCodes::HTTP_BAD_REQUEST
-            );
-            $hook = $hookBase . '_FAIL';
-            $msg = json_encode(
-                [
-                    'error' => $e->getMessage(),
-                    'title' => $failTitle
-                ]
-            );
-        }
-        $this->jsonHookResponse(
-            [
-                $entityKey => &$Entity,
-                'hook' => &$hook,
-                'code' => &$code,
-                'msg' => &$msg,
-                'serverFault' => &$serverFault
-            ],
-            $hook
-        );
-    }
-    /**
-     * Shared scaffold for the update (editPost) AJAX handlers.
-     *
-     * The edit counterpart of handleAddPost(). It owns the same
-     * boilerplate, with three differences inherent to editing an
-     * existing entity: the "<BASE>_POST" pre-event carries the loaded
-     * entity ([$entityKey => &$this->obj]) so listeners can inspect or
-     * replace it; the success status is HTTP 202 Accepted rather than
-     * 201 Created; and the entity is the page's own $this->obj, so the
-     * $build closure mutates it in place and need not return anything.
-     *
-     * The page-specific part — reading $_POST, applying changes to
-     * $this->obj, and saving — lives in the $build closure. The closure
-     * receives $serverFault by reference (set it true before throwing to
-     * signal an HTTP 500 rather than a 400). $this->obj is handed to the
-     * result hook under $entityKey so listeners on "<BASE>_SUCCESS" /
-     * "<BASE>_FAIL" still see it exactly as before.
-     *
-     * @param string   $entityKey    Payload key for the entity (e.g. 'Group').
-     * @param string   $hookBase     Hook prefix (e.g. 'GROUP_EDIT'); the
-     *                               _POST/_SUCCESS/_FAIL events derive from it.
-     * @param string   $successMsg   Translated success message body.
-     * @param string   $successTitle Translated success title.
-     * @param string   $failTitle    Translated failure title.
-     * @param callable $build        Closure(&$serverFault): mutates $this->obj.
-     *
-     * @return void
-     */
-    protected function handleEditPost(
-        $entityKey,
-        $hookBase,
-        $successMsg,
-        $successTitle,
-        $failTitle,
-        callable $build
-    ) {
-        self::checkAuthAndCSRF();
-        header('Content-type: application/json');
-        self::$HookManager->processEvent(
-            $hookBase . '_POST',
-            [$entityKey => &$this->obj]
-        );
-        $serverFault = false;
-        try {
-            $build($serverFault);
-            $code = HTTPResponseCodes::HTTP_ACCEPTED;
-            $hook = $hookBase . '_SUCCESS';
-            $msg = json_encode(
-                [
-                    'msg' => $successMsg,
-                    'title' => $successTitle
-                ]
-            );
-        } catch (Exception $e) {
-            $code = (
-                $serverFault ?
-                HTTPResponseCodes::HTTP_INTERNAL_SERVER_ERROR :
-                HTTPResponseCodes::HTTP_BAD_REQUEST
-            );
-            $hook = $hookBase . '_FAIL';
-            $msg = json_encode(
-                [
-                    'error' => $e->getMessage(),
-                    'title' => $failTitle
-                ]
-            );
-        }
-        $this->jsonHookResponse(
-            [
-                $entityKey => &$this->obj,
-                'hook' => &$hook,
-                'code' => &$code,
-                'msg' => &$msg,
-                'serverFault' => &$serverFault
-            ],
-            $hook
-        );
-    }
-    /**
      * Actually performs the deletion of selected items.
      *
      * @return void
@@ -2446,288 +2251,6 @@ abstract class FOGPage extends FOGBase
             ),
             '',
             'warning'
-        );
-    }
-    /**
-     * Renders a standard association tab: a primary box containing the
-     * add/remove buttons, the server-side list table, and the dissociate
-     * confirmation modal.
-     *
-     * @param string $tabSlug   node-sub slug (e.g. 'host-group') driving the
-     *                          button ids, table id, and tab update URL
-     * @param string $boxTitle  translated box title (e.g. _('Host Group Associations'))
-     * @param string $colHeader translated first-column header (e.g. _('Group Name'))
-     * @param string $delItem   singular item name passed to assocDelModal (e.g. 'group')
-     * @param string $sendClass css class for the "Add selected" button (some tabs
-     *                          use 'btn btn-success pull-right' instead of primary)
-     * @param string $helpBlock optional translated help text rendered as a
-     *                          help-block in the box header (already escaped/safe)
-     *
-     * @return void
-     */
-    protected function renderAssocTab(
-        $tabSlug,
-        $boxTitle,
-        $colHeader,
-        $delItem,
-        $sendClass = 'btn btn-primary pull-right',
-        $helpBlock = ''
-    ) {
-        $this->headerData = [
-            $colHeader,
-            _('Associated')
-        ];
-        $this->attributes = [
-            [],
-            ['width' => 16]
-        ];
-        $props = ' method="post" action="'
-            . self::makeTabUpdateURL(
-                $tabSlug,
-                $this->obj->get('id')
-            )
-            . '" ';
-
-        $buttons = self::makeButton(
-            "$tabSlug-send",
-            _('Add selected'),
-            $sendClass,
-            $props
-        );
-        $buttons .= self::makeButton(
-            "$tabSlug-remove",
-            _('Remove selected'),
-            'btn btn-danger pull-left',
-            $props
-        );
-
-        echo '<div class="box box-primary">';
-        echo '<div class="box-header with-border">';
-        echo '<h4 class="box-title">';
-        echo $boxTitle;
-        echo '</h4>';
-        if ($helpBlock !== '') {
-            echo '<p class="help-block">';
-            echo $helpBlock;
-            echo '</p>';
-        }
-        echo '</div>';
-        echo '<div class="box-body">';
-        $this->render(12, "$tabSlug-table", $buttons);
-        echo '</div>';
-        echo '<div class="box-footer with-border">';
-        echo $this->assocDelModal($delItem);
-        echo '</div>';
-        echo '</div>';
-    }
-    /**
-     * Handles a standard association add/remove POST: reads the additems /
-     * remitems arrays and dispatches them to the object's add/remove methods.
-     * When $orderMethod is supplied, also honours a snapinorder array (used by
-     * the group/host snapin tabs to persist execution order).
-     *
-     * @param string $addMethod    obj method to add associations (e.g. 'addGroup')
-     * @param string $removeMethod obj method to remove associations (e.g. 'removeGroup')
-     * @param string $orderMethod  obj method to set ordering from the snapinorder
-     *                             POST array, or null when the tab has no ordering
-     *
-     * @return void
-     */
-    protected function assocPost($addMethod, $removeMethod, $orderMethod = null)
-    {
-        self::checkAuthAndCSRF();
-        if (isset($_POST['confirmadd'])) {
-            $items = filter_input_array(
-                INPUT_POST,
-                [
-                    'additems' => [
-                        'flags' => FILTER_REQUIRE_ARRAY
-                    ]
-                ]
-            );
-            $items = $items['additems'];
-            if (count($items ?: []) > 0) {
-                $this->obj->{$addMethod}($items);
-            }
-        }
-        if (isset($_POST['confirmdel'])) {
-            $items = filter_input_array(
-                INPUT_POST,
-                [
-                    'remitems' => [
-                        'flags' => FILTER_REQUIRE_ARRAY
-                    ]
-                ]
-            );
-            $items = $items['remitems'];
-            if (count($items ?: []) > 0) {
-                $this->obj->{$removeMethod}($items);
-            }
-        }
-        if ($orderMethod !== null && isset($_POST['snapinorder'])) {
-            $order = filter_input_array(
-                INPUT_POST,
-                [
-                    'snapinorder' => [
-                        'flags' => FILTER_REQUIRE_ARRAY
-                    ]
-                ]
-            );
-            $order = $order['snapinorder'];
-            if (count($order ?: []) > 0) {
-                $this->obj->{$orderMethod}($order);
-            }
-        }
-    }
-    /**
-     * Renders a simple display tab: a box-primary panel with a title and a
-     * single DataTable. Shared by the group/host history tabs whose only
-     * differences are the column set, the title text and the table id.
-     *
-     * @param array  $headerData The column headers (already translated).
-     * @param array  $attributes The per-column attribute arrays.
-     * @param string $title      The box title (already translated).
-     * @param string $tableId    The DataTable element id.
-     *
-     * @return void
-     */
-    protected function renderHistoryTab(array $headerData, array $attributes, $title, $tableId)
-    {
-        $this->headerData = $headerData;
-        $this->attributes = $attributes;
-        echo '<div class="box box-primary">';
-        echo '<div class="box-header with-border">';
-        echo '<h4 class="box-title">';
-        echo $title;
-        echo '</h4>';
-        echo '</div>';
-        echo '<div class="box-body">';
-        $this->render(12, $tableId);
-        echo '</div>';
-        echo '</div>';
-    }
-    /**
-     * Streams a Login/Image history datatable payload as JSON.
-     *
-     * Shared by the host and group Login/Image history AJAX endpoints, which
-     * differ only in the host scope (a single host id for a host, the member
-     * host ids for a group) and the route resource to list.
-     *
-     * @param mixed  $scope the hostID scope (int for a host, array for a group)
-     * @param string $route the Route resource to list (e.g. 'usertracking')
-     *
-     * @return void
-     */
-    protected function renderHistoryData($scope, $route)
-    {
-        header('Content-type: application/json');
-        parse_str(
-            file_get_contents('php://input'),
-            $pass_vars
-        );
-
-        Route::listem(
-            $route,
-            ['hostID' => $scope]
-        );
-        echo Route::getData();
-        exit;
-    }
-    /**
-     * Streams the snapin-task history datatable payload as JSON.
-     *
-     * Shared by the host and group Snapin history AJAX endpoints; differs only
-     * in the host scope. Returns an empty datatable payload (rather than an
-     * unscoped lookup) when the scope has no snapin jobs.
-     *
-     * @param mixed $scope the hostID scope (int for a host, array for a group)
-     *
-     * @return void
-     */
-    protected function renderSnapinHistoryData($scope)
-    {
-        header('Content-type: application/json');
-        parse_str(
-            file_get_contents('php://input'),
-            $pass_vars
-        );
-
-        $checkStates = [
-            self::getCancelledState(),
-            self::getCompleteState()
-        ];
-
-        $snapinJobs = Route::getIds(
-            'snapinjob',
-            ['hostID' => $scope]
-        );
-        $snapinJobs = array_filter(
-            array_map('intval', (array)$snapinJobs),
-            function ($id) {
-                return $id > 0;
-            }
-        );
-
-        // If there are no jobs in scope, return an empty datatable payload and
-        // avoid an unscoped snapintask lookup.
-        if (count($snapinJobs) < 1) {
-            $this->jsonSend(HTTPResponseCodes::HTTP_SUCCESS, json_encode(
-                [
-                    'draw' => (int)filter_input(INPUT_POST, 'draw') ?: 0,
-                    'recordsTotal' => 0,
-                    'recordsFiltered' => 0,
-                    'data' => [],
-                    '_lang' => 'snapintask'
-                ]
-            ));
-        }
-
-        Route::listem(
-            'snapintask',
-            [
-                'jobID' => $snapinJobs,
-                'stateID' => $checkStates
-            ]
-        );
-
-        echo Route::getData();
-        exit;
-    }
-    /**
-     * Builds a standard association list table via getItemsList using a LEFT
-     * OUTER JOIN that flags which rows are already associated with the current
-     * object.
-     *
-     * @param string $itemType     item class to list (e.g. 'group')
-     * @param string $listType     list/association type key (e.g. 'groupassociation')
-     * @param string $assocTable   join table (e.g. 'groupMembers')
-     * @param string $itemKey      listed item's key column (e.g. '`groups`.`groupID`')
-     * @param string $assocItemKey join table's item column (e.g. '`groupMembers`.`gmGroupID`')
-     * @param string $ownerKey     join table's owner column (e.g. '`groupMembers`.`gmHostID`')
-     * @param array  $columns      extra association column definition(s)
-     *
-     * @return void
-     */
-    protected function assocItemsList(
-        $itemType,
-        $listType,
-        $assocTable,
-        $itemKey,
-        $assocItemKey,
-        $ownerKey,
-        array $columns
-    ) {
-        $join = [
-            "LEFT OUTER JOIN `$assocTable` ON "
-            . "$itemKey = $assocItemKey "
-            . "AND $ownerKey = '" . $this->obj->get('id') . "'"
-        ];
-        return $this->obj->getItemsList(
-            $itemType,
-            $listType,
-            $join,
-            '',
-            $columns
         );
     }
     /**
@@ -4912,38 +4435,6 @@ abstract class FOGPage extends FOGBase
         return ob_get_clean();
     }
     /**
-     * Shared scaffold for the standard edit() tab pages.
-     *
-     * Sets the canonical "Edit: <name> ID: <id>" page title from the
-     * loaded $this->obj and echoes the assembled tab markup. The page's
-     * own edit() keeps building its page-specific $tabData entries and
-     * hands the finished array here; the title and the tabFields() echo
-     * are the only shared bookends, so they live here.
-     *
-     * The $obj argument is passed straight through to tabFields() with
-     * the same -1 default, preserving its three behaviours: -1 rebuilds
-     * the entity from the node/id globals, an explicit object uses it as
-     * given, and false skips the TABDATA/plugin hook injection. The page
-     * title always derives from $this->obj regardless of $obj.
-     *
-     * @param array $tabData The page's assembled tab definitions.
-     * @param mixed $obj     tabFields() obj arg: -1 (default), an entity,
-     *                       or false. Does not affect the title.
-     *
-     * @return void
-     */
-    protected function renderEditTabs(array $tabData, $obj = -1)
-    {
-        $this->title = sprintf(
-            '%s: %s %s: %s',
-            _('Edit'),
-            $this->obj->get('name'),
-            _('ID'),
-            $this->obj->get('id')
-        );
-        echo self::tabFields($tabData, $obj);
-    }
-    /**
      * Function passes so we can have
      * a paged version of universal searching.
      *
@@ -5059,386 +4550,6 @@ abstract class FOGPage extends FOGBase
             . " title='$title'"
             . ($extra ? " $extra" : '')
             . '></i>';
-    }
-    /**
-     * Renders the standard create-form scaffold shared by the add() pages:
-     * a CSRF-protected horizontal form wrapping a box-solid that holds one or
-     * more titled box-primary sections, with the action buttons in the footer.
-     *
-     * @param string $idBase   Id prefix; yields form id "<idBase>-create-form"
-     *                         and container id "<idBase>-create".
-     * @param array  $sections List of [title, body] pairs; each renders as a
-     *                         box-primary (one for most pages, two for hosts).
-     * @param string $buttons  Pre-rendered footer buttons.
-     * @param string $enctype  Form enctype (default urlencoded; snapin uses
-     *                         multipart/form-data).
-     *
-     * @return void
-     */
-    protected function renderCreateForm(
-        $idBase,
-        array $sections,
-        $buttons,
-        $enctype = 'application/x-www-form-urlencoded'
-    ) {
-        echo self::makeFormTag(
-            'form-horizontal',
-            $idBase . '-create-form',
-            $this->formAction,
-            'post',
-            $enctype,
-            true
-        );
-        echo '<div class="box box-solid" id="' . $idBase . '-create">';
-        echo '<div class="box-body">';
-        foreach ($sections as $section) {
-            echo '<div class="box box-primary">';
-            echo '<div class="box-header with-border">';
-            echo '<h4 class="box-title">';
-            echo $section[0];
-            echo '</h4>';
-            echo '</div>';
-            echo '<div class="box-body">';
-            echo $section[1];
-            echo '</div>';
-            echo '</div>';
-        }
-        echo '</div>';
-        echo '<div class="box-footer with-border">';
-        echo $buttons;
-        echo '</div>';
-        echo '</div>';
-        echo '</form>';
-    }
-    /**
-     * Renders the standard "Create New X" page form.
-     *
-     * Wraps the near-identical add() body shared by nearly every management
-     * page: set the page title, build the create-form fields via _addFields(),
-     * add the uniform Create button, fire the page's *_ADD_FIELDS hook (with
-     * the fields, buttons, and the entity class in the payload), then hand the
-     * single titled section off to renderCreateForm().
-     *
-     * The section title shown above the fields is the same text as the page
-     * title, matching every page that used this template.
-     *
-     * @param string      $idBase      renderCreateForm id base (e.g. 'group')
-     * @param string      $title       page + section title (already _()'d)
-     * @param string      $hookEvent   the *_ADD_FIELDS event name to fire
-     * @param string      $entityKey   payload key for the entity class
-     * @param string|null $entityClass class to instantiate (defaults to key)
-     * @param string      $enctype     form enctype, default urlencoded
-     *
-     * @return void
-     */
-    protected function renderAddForm(
-        $idBase,
-        $title,
-        $hookEvent,
-        $entityKey,
-        $entityClass = null,
-        $enctype = 'application/x-www-form-urlencoded'
-    ) {
-        $this->title = $title;
-
-        $fields = $this->_addFields();
-
-        $buttons = self::makeButton(
-            'send',
-            _('Create'),
-            'btn btn-primary pull-right'
-        );
-
-        self::$HookManager->processEvent(
-            $hookEvent,
-            [
-                'fields' => &$fields,
-                'buttons' => &$buttons,
-                $entityKey => self::getClass($entityClass ?? $entityKey)
-            ]
-        );
-        $rendered = self::formFields($fields);
-        unset($fields);
-
-        $this->renderCreateForm(
-            $idBase,
-            [[$title, $rendered]],
-            $buttons,
-            $enctype
-        );
-    }
-    /**
-     * Renders the standard create form fragment used inside the "add" modal.
-     *
-     * Wraps the near-identical addModal() body shared by nearly every
-     * management page: build the create-form fields via _addFields(), fire the
-     * page's *_ADD_FIELDS hook (fields + entity class, no buttons), then echo a
-     * bare form tag, the rendered fields, and the closing tag.
-     *
-     * @param string      $node        URL node for the form action target
-     * @param string      $hookEvent   the *_ADD_FIELDS event name to fire
-     * @param string      $entityKey   payload key for the entity class
-     * @param string|null $entityClass class to instantiate (defaults to key)
-     * @param string      $enctype     form enctype, default urlencoded
-     *
-     * @return void
-     */
-    protected function renderAddModalForm(
-        $node,
-        $hookEvent,
-        $entityKey,
-        $entityClass = null,
-        $enctype = 'application/x-www-form-urlencoded'
-    ) {
-        $fields = $this->_addFields();
-
-        self::$HookManager->processEvent(
-            $hookEvent,
-            [
-                'fields' => &$fields,
-                $entityKey => self::getClass($entityClass ?? $entityKey)
-            ]
-        );
-        $rendered = self::formFields($fields);
-        unset($fields);
-
-        echo self::makeFormTag(
-            'form-horizontal',
-            'create-form',
-            '../management/index.php?node=' . $node . '&sub=add',
-            'post',
-            $enctype,
-            true
-        );
-        echo $rendered;
-        echo '</form>';
-    }
-    /**
-     * Builds the schedule-type form fields shared by the host/group deploy()
-     * create-task forms: the always-present "Schedule Immediately" radio plus,
-     * unless this is a debug or password-reset task, the "Schedule Later"
-     * (single) and "Schedule Crontab Style" (cron) inputs.
-     *
-     * @param string $labelClass The label class used by the deploy form.
-     * @param bool   $isdebug    Whether this is a debug-session task.
-     * @param int    $type       The task type id (to suppress for resets).
-     *
-     * @return array The field fragment to fastmerge onto the form fields.
-     */
-    protected function scheduleTypeFields($labelClass, $isdebug, $type)
-    {
-        $fields = [
-            self::makeLabel(
-                $labelClass,
-                'instant',
-                _('Schedule Immediately')
-            ) => self::makeInput(
-                'instant',
-                'scheduleType',
-                '',
-                'radio',
-                'instant',
-                'instant',
-                false,
-                false,
-                -1,
-                -1,
-                ' checked'
-            )
-        ];
-        if (!$isdebug
-            && TaskType::PASSWORD_RESET != $type
-        ) {
-            $fields = self::fastmerge(
-                $fields,
-                [
-                    '<div class="hideFromDebug">'
-                    . self::makeLabel(
-                        $labelClass,
-                        'delayed',
-                        _('Schedule Later')
-                    ) => self::makeInput(
-                        'delayed',
-                        'scheduleType',
-                        '',
-                        'radio',
-                        'delayed',
-                        'single'
-                    )
-                    . '</div>',
-                    '<div class="delayedinput hidden">'
-                    . self::makeLabel(
-                        $labelClass,
-                        'delayedinput',
-                        _('Start Time')
-                    ) => self::makeInput(
-                        'form-control',
-                        'scheduleSingleTime',
-                        self::niceDate()->format('Y-m-d H:i:s'),
-                        'text',
-                        'delayedinput',
-                        ''
-                    )
-                    . '</div>',
-                    '<div class="hideFromDebug">'
-                    . self::makeLabel(
-                        $labelClass,
-                        'cron',
-                        _('Schedule Crontab Style')
-                    ) => self::makeInput(
-                        'croninput',
-                        'scheduleType',
-                        '',
-                        'radio',
-                        'cron',
-                        'cron'
-                    )
-                    . '</div>',
-                    '<div class="croninput hidden">'
-                    . self::makeLabel(
-                        $labelClass,
-                        '',
-                        _('Cron Entry')
-                    ) => '<div class="croninput fogcron hidden"></div><br/>'
-                    . self::makeInput(
-                        'col-sm-2 croninput cronmin hidden',
-                        'scheduleCronMin',
-                        _('min'),
-                        'text',
-                        'cronMin'
-                    )
-                    . self::makeInput(
-                        'col-sm-2 croninput cronhour hidden',
-                        'scheduleCronHour',
-                        _('hour'),
-                        'text',
-                        'cronHour'
-                    )
-                    . self::makeInput(
-                        'col-sm-2 croninput crondom hidden',
-                        'scheduleCronDOM',
-                        _('day'),
-                        'text',
-                        'cronDom'
-                    )
-                    . self::makeInput(
-                        'col-sm-2 croninput cronmonth hidden',
-                        'scheduleCronMonth',
-                        _('month'),
-                        'text',
-                        'cronMonth'
-                    )
-                    . self::makeInput(
-                        'col-sm-2 croninput crondow hidden',
-                        'scheduleCronDOW',
-                        _('weekday'),
-                        'text',
-                        'cronDow'
-                    )
-                    . '</div>'
-                ]
-            );
-        }
-        return $fields;
-    }
-    /**
-     * Validates the posted schedule type for the host/group deploy() create-task
-     * handlers and resolves its parameters. Honors the SCHEDULE_TYPES hook,
-     * rejects an unknown type, a past single-run time, or an invalid cron field.
-     *
-     * @throws Exception If the schedule type or any cron/time field is invalid.
-     *
-     * @return array Keyed: scheduleType, scheduleDeployTime (single), and
-     *               min/hour/dom/month/dow (cron); unused entries are null.
-     */
-    protected function validateScheduleType()
-    {
-        $scheduleType = strtolower(
-            filter_input(INPUT_POST, 'scheduleType')
-        );
-        $scheduleTypes = [
-            'cron',
-            'instant',
-            'single'
-        ];
-        self::$HookManager->processEvent(
-            'SCHEDULE_TYPES',
-            ['scheduleTypes' => &$scheduleTypes]
-        );
-        foreach ($scheduleTypes as $ind => &$val) {
-            $scheduleTypes[$ind] = trim(
-                strtolower(
-                    $val
-                )
-            );
-            unset($val);
-        }
-        if (!in_array($scheduleType, $scheduleTypes)) {
-            throw new Exception(_('Invalid scheduling type'));
-        }
-        $schedule = [
-            'scheduleType' => $scheduleType,
-            'scheduleDeployTime' => null,
-            'min' => null,
-            'hour' => null,
-            'dom' => null,
-            'month' => null,
-            'dow' => null
-        ];
-        // Schedule Delayed/Cron checks.
-        switch ($scheduleType) {
-            case 'single':
-                $scheduleDeployTime = self::niceDate(
-                    filter_input(INPUT_POST, 'scheduleSingleTime')
-                );
-                if ($scheduleDeployTime < self::niceDate()) {
-                    throw new Exception(_('Scheduled time is in the past'));
-                }
-                $schedule['scheduleDeployTime'] = $scheduleDeployTime;
-                break;
-            case 'cron':
-                $min = strval(
-                    filter_input(INPUT_POST, 'scheduleCronMin')
-                );
-                $hour = strval(
-                    filter_input(INPUT_POST, 'scheduleCronHour')
-                );
-                $dom = strval(
-                    filter_input(INPUT_POST, 'scheduleCronDOM')
-                );
-                $month = strval(
-                    filter_input(INPUT_POST, 'scheduleCronMonth')
-                );
-                $dow = strval(
-                    filter_input(INPUT_POST, 'scheduleCronDOW')
-                );
-                $tmin = FOGCron::checkMinutesField($min);
-                $thour = FOGCron::checkHoursField($hour);
-                $tdom = FOGCron::checkDOMField($dom);
-                $tmonth = FOGCron::checkMonthField($month);
-                $tdow = FOGCron::checkDOWField($dow);
-                if (!$tmin) {
-                    throw new Exception(_('Minutes field is invalid'));
-                }
-                if (!$thour) {
-                    throw new Exception(_('Hours field is invalid'));
-                }
-                if (!$tdom) {
-                    throw new Exception(_('Day of Month field is invalid'));
-                }
-                if (!$tmonth) {
-                    throw new Exception(_('Month field is invalid'));
-                }
-                if (!$tdow) {
-                    throw new Exception(_('Day of Week field is invalid'));
-                }
-                $schedule['min'] = $min;
-                $schedule['hour'] = $hour;
-                $schedule['dom'] = $dom;
-                $schedule['month'] = $month;
-                $schedule['dow'] = $dow;
-        }
-        return $schedule;
     }
     /**
      * Makes the opening form tag.
