+++ packages/web/lib/fog/system.class.php
@@ -59,7 +59,7 @@ class System
     public function __construct()
     {
         self::_versionCompare();
-        define('FOG_VERSION', '1.6.0-beta.2505');
+        define('FOG_VERSION', '1.6.0-beta.2506');
         define('FOG_CHANNEL', 'Beta');
         define('FOG_SCHEMA', 298);
         define('FOG_BCACHE_VER', 173);
